refactor(process): added new sub module in process for more readability
